Refactored dice rolling in sheet, moved styles to scss. Moved rolling code to dice-helpers.js, and updated reference for actor and adversary sheets. Added die rolling by clicking on the die icon over the chat window.  It will open a blank rolling dialog for ad-hoc rolling.
