+++ scss/components/_dicepool.scss
@@ -1,0 +1,31 @@
+.dice-pool-dialog {
+  .dice-pool img {
+    border: 0;
+  }
+  .dice-pool {
+    min-height: 51px;
+    margin: auto;
+  }
+  .pool-container img {
+    border: 0;
+  }
+  .pool-value {
+    height: 42px;
+    border: 1px solid black;
+    margin: auto;
+    position: relative;
+    background-color: #ddd;
+  }
+  .pool-value input {
+    width: 100%;
+    height: 100%;
+    border: 0;
+    font-size: 30px;
+    text-align: center;
+    background: none;
+  }
+  .upgrade-buttons {
+    display: flex;
+    margin: 10px;
+  }
+}
