+++ scss/starwarsffg.scss
@@ -9,6 +9,8 @@
 @import "global/grid";
 @import "global/flex";
 
+
+
 .starwarsffg {
   @import "components/forms";
   @import "components/sheet";
@@ -37,5 +39,6 @@
   @import 'components/forcepowers';
   @import 'components/editor';
   @import 'components/specializations';
-  @import 'components/dataimporter'
+  @import 'components/dataimporter';
+  @import 'components/dicepool';
 }
