+++ templates/roll-options.html
@@ -1,34 +1,4 @@
-<div id="{{id}}">
-  <style>
-    .dice-pool img {
-      border: 0;
-    }
-    .dice-pool {
-      margin: auto;
-    }
-    .pool-container img {
-      border: 0;
-    }
-    .pool-value {
-      height: 42px;
-      border: 1px solid black;
-      margin: auto;
-      position: relative;
-      background-color: #ddd;
-    }
-    .pool-value input {
-      width: 100%;
-      height: 100%;
-      border: 0;
-      font-size: 30px;
-      text-align: center;
-      background: none;
-    }
-    .upgrade-buttons {
-      display: flex;
-      margin: 10px;
-    }
-  </style>
+<div id="{{id}}" class="dice-pool-dialog">
   <div class="dice-pool"></div>
   <div class="upgrade-buttons">
     <button id="upgrade-ability">{{localize "SWFFG.UpgradeAbility"}}</button>
